backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
werkzeug==3.0.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==2.0.3
PyPDF2==3.0.1
pdfplumber==0.10.3
SpeechRecognition==3.10.0
requests==2.31.0
python-dotenv==1.0.0